fixed: Refactored market page structure and Redux logic
