+++ src/store/slices/Market/thunks.ts
@@ -1,0 +1,19 @@
+import { createAsyncThunk } from "@reduxjs/toolkit";
+import API from "@api/index";
+
+export const fetchProducts = createAsyncThunk(
+    "market/fetchProducts",
+    async () => {
+        try {
+            const response = await API.getProducts();
+
+            return response;
+        } catch (error) {
+            if (error instanceof Error) {
+                throw new Error(error.message);
+            }
+
+            throw new Error("An unknown error occurred");
+        }
+    }
+);
